Move browser locale logic to userSettingsContext
