+++ app/i18n/Provider.tsx
@@ -1,10 +1,8 @@
 "use client";
-import { useEffect } from "react";
 import { Fragment } from "react";
 import { IntlProvider } from "react-intl";
 import { useUserSettingsContext } from "@contexts/UserSettingsContext";
 
-import { LOCALES } from "@contexts/UserSettingsContext";
 import messages from "./messages";
 
 type Props = {
@@ -12,18 +10,8 @@ type Props = {
 };
 
 const Provider = ({ children }: Props) => {
-  const { userSettings, dispatch } = useUserSettingsContext();
+  const { userSettings } = useUserSettingsContext();
   const { locale } = userSettings;
-
-  useEffect(() => {
-    const locale = navigator.language;
-    const baseLocale = locale.split("-")[0];
-    if (baseLocale === "en") {
-      dispatch({ type: "SET_LOCALE", locale: LOCALES.ENGLISH });
-    } else if (baseLocale === "ja") {
-      dispatch({ type: "SET_LOCALE", locale: LOCALES.JAPANESE });
-    }
-  }, [dispatch]);
 
   return (
     <IntlProvider
